Workload schema fixes
