+++ pkg/api/server/userstored/setup.go
@@ -30,19 +30,19 @@ func Setup(ctx context.Context, mgmt *config.ScaledContext, clusterManager *clus
 	schemas := mgmt.Schemas
 
 	addProxyStore(ctx, schemas, mgmt, client.ConfigMapType, "v1", nil)
-	addProxyStore(ctx, schemas, mgmt, client.CronJobType, "batch/v1beta1", workload.New)
-	addProxyStore(ctx, schemas, mgmt, client.DaemonSetType, "apps/v1beta2", workload.New)
-	addProxyStore(ctx, schemas, mgmt, client.DeploymentType, "apps/v1beta2", workload.New)
+	addProxyStore(ctx, schemas, mgmt, client.CronJobType, "batch/v1beta1", workload.NewCustomizeStore)
+	addProxyStore(ctx, schemas, mgmt, client.DaemonSetType, "apps/v1beta2", workload.NewCustomizeStore)
+	addProxyStore(ctx, schemas, mgmt, client.DeploymentType, "apps/v1beta2", workload.NewCustomizeStore)
 	addProxyStore(ctx, schemas, mgmt, client.IngressType, "extensions/v1beta1", ingress.Wrap)
-	addProxyStore(ctx, schemas, mgmt, client.JobType, "batch/v1", workload.New)
+	addProxyStore(ctx, schemas, mgmt, client.JobType, "batch/v1", workload.NewCustomizeStore)
 	addProxyStore(ctx, schemas, mgmt, client.PersistentVolumeClaimType, "v1", nil)
 	addProxyStore(ctx, schemas, mgmt, client.PodType, "v1", func(store types.Store) types.Store {
 		return pod.New(store, clusterManager, mgmt)
 	})
-	addProxyStore(ctx, schemas, mgmt, client.ReplicaSetType, "apps/v1beta2", workload.New)
-	addProxyStore(ctx, schemas, mgmt, client.ReplicationControllerType, "v1", workload.New)
+	addProxyStore(ctx, schemas, mgmt, client.ReplicaSetType, "apps/v1beta2", workload.NewCustomizeStore)
+	addProxyStore(ctx, schemas, mgmt, client.ReplicationControllerType, "v1", workload.NewCustomizeStore)
 	addProxyStore(ctx, schemas, mgmt, client.ServiceType, "v1", service.New)
-	addProxyStore(ctx, schemas, mgmt, client.StatefulSetType, "apps/v1beta2", workload.New)
+	addProxyStore(ctx, schemas, mgmt, client.StatefulSetType, "apps/v1beta2", workload.NewCustomizeStore)
 	addProxyStore(ctx, schemas, mgmt, clusterClient.NamespaceType, "v1", namespace.New)
 	addProxyStore(ctx, schemas, mgmt, clusterClient.PersistentVolumeType, "v1", nil)
 	addProxyStore(ctx, schemas, mgmt, clusterClient.StorageClassType, "storage.k8s.io/v1", nil)
@@ -88,7 +88,7 @@ func Namespace(schemas *types.Schemas, manager *clustermanager.Manager) {
 }
 
 func Workload(schemas *types.Schemas, clusterManager *clustermanager.Manager) {
-	workload.ConfigureStore(schemas, clusterManager)
+	workload.NewWorkloadAggregateStore(schemas, clusterManager)
 }
 
 func Service(schemas *types.Schemas) {
